+++ src/client/src/commons/form/borehole/completion/backfillDisplay.jsx
@@ -1,6 +1,7 @@
 import { StackFullWidth } from "../../../../components/baseComponents";
 import { FormDisplay, FormDisplayType } from "../../../../components/form/form";
 import DataDisplayCard from "../../../../components/dataCard/dataDisplayCard";
+import { StackHalfWidth } from "../../../../components/baseComponents.js";
 
 const BackfillDisplay = props => {
   const { item, selected, setSelected, isEditable, deleteData } = props;
@@ -20,6 +21,9 @@ const BackfillDisplay = props => {
         <FormDisplay label="kindFilling" value={item?.kind} type={FormDisplayType.Domain} />
         <FormDisplay label="materialFilling" value={item?.material} type={FormDisplayType.Domain} />
       </StackFullWidth>
+      <StackHalfWidth>
+        <FormDisplay label="casingName" value={item?.casingId ? item?.casing?.name : null} />
+      </StackHalfWidth>
       <FormDisplay label="notes" value={item?.notes} />
     </DataDisplayCard>
   );